Module Bench.Quota

A quota can be specified as an amount of wall time, or a number of times to run the function.

(Strictly speaking, for Num_calls n, it is possible that the function is called fewer than n times if the array of measurements fills up. But with default settings for how batches are sized, you don't run into this issue until n is over 1.9e16.)

type t =

| Span of Core.Time.Span.t

| Num_calls of int

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

Examples:

"10" -> Span 10s (float-like: convert to seconds)

"1m" -> Span 1m (span-like: keep as span)

"5x" -> Num_calls 5

"1e9x" -> Num_calls 1_000_000_000

include Core.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

val arg_type : t Core.Command.Param.Arg_type.t

val fulfilled : t -> start:Core.Time.t -> num_calls:int -> bool

fulfilled t ~start ~num_calls returns true iff we have fulfilled the quota, given that we *started* at time start and have run the function num_calls times.

val max_count : t -> int

max_count (Num_calls n) returns n, max_count (Span _) returns max_int.

val scale_int : t -> int -> t

Scale by an integer factor