Module Core_bench_js

module Quota = Core_bench_internals.Quota

module Test = Core_bench_internals.Test

module Run_config = Core_bench_internals.Run_config

module Display_column = Core_bench_internals.Display_column

module Display_config = Core_bench_internals.Display_config

module Variable = Core_bench_internals.Variable

module Measurement = Core_bench_internals.Measurement

module Analysis_config = Core_bench_internals.Analysis_config

module Analysis_result = Core_bench_internals.Analysis_result

val bench : ?run_config:Run_config.t -> ?analysis_configs:Analysis_config.t list -> ?display_config:Display_config.t -> ?save_to_file:( Measurement.t -> string ) -> ?libname:string -> Test.t list -> unit

bench tests will run, analyze and display the specified tests. Use this when one needs more control over the execution parameters that what is exposed through make_command. bench can also save the measurements of each test to the filename returned by save_to_file.

val measure : ?run_config:Run_config.t -> Test.t list -> Measurement.t list

measure is a fragment of the functionality of bench. measure tests will run the specified tests and return the resulting measurement results.

val analyze : ?analysis_configs:Analysis_config.t list -> Measurement.t -> Analysis_result.t Core.Or_error.t

analyze is a fragment of the functionality of bench. analyze ~analysis_configs m will analyze the measurement m using the regressions specified.

val display : ?libname:string -> ?display_config:Display_config.t -> Analysis_result.t list -> unit

display is a fragment of the functionality of bench. display results will display a tabular summary of results on the terminal.